refactor: split accumulated.py into public API and internal helpers
Extract data-loading logic into _accumulated_data.py (59 lines) and
computation helpers into _accumulated_helpers.py (98 lines), keeping
accumulated.py (96 lines) as the public API with re-exports for
backward compatibility.
